Decouple tiles from occupants, scale NPC sandbox to 10,000 NPCs
Tiles are now pure terrain (all 8 bits = type). Occupancy tracked
in separate OccGrid []uint16 with O(1) npcByID map. NPC.ID widened
from byte to uint16 (65535 max, no wrapping). All Nearest* functions
use bounded Manhattan ring search (radius 0-31) instead of full grid
scan. Cached foodCount/itemCount maintained by SetTile(). Auto-scale
world size to ~6% density. Trades and teaches occur at all scales.

+++ pkg/sandbox/npc.go
@@ -108,7 +108,7 @@ var ItemModifiers = map[byte]Modifier{
 
 // NPC represents a creature in the sandbox world.
 type NPC struct {
-	ID      byte
+	ID      uint16
 	X, Y    int
 	Health  int
 	Energy  int
